move away buffer
